feat: made some updates

+++ nebula-app/src/app/api/controllers/category.controller.ts
@@ -1,6 +1,7 @@
 import { NextRequest } from "next/server";
 import { CategoryService } from "../services/category.service";
 import { createCategorySchema, updateCategorySchema } from "@/lib/validations";
+import { BadRequestException } from "../utils/http-exception";
 
 export class CategoryController {
   async getAll(request: NextRequest) {
@@ -8,6 +9,10 @@ export class CategoryController {
   }
 
   async getById(request: NextRequest, id: string) {
+    if (!id) {
+      throw new BadRequestException("Category ID is required");
+    }
+
     return await CategoryService.getById(id);
   }
 
@@ -18,12 +23,20 @@ export class CategoryController {
   }
 
   async update(request: NextRequest, id: string) {
+    if (!id) {
+      throw new BadRequestException("Category ID is required");
+    }
+
     const body = await request.json();
     const payload = updateCategorySchema.parse(body);
     return await CategoryService.update(id, payload);
   }
 
   async delete(request: NextRequest, id: string) {
+    if (!id) {
+      throw new BadRequestException("Category ID is required");
+    }
+
     return await CategoryService.delete(id);
   }
 }
