+++ nebula-app/src/app/api/controllers/event.controller.ts
@@ -1,24 +1,21 @@
-import { NextRequest, NextResponse } from "next/server";
+import { NextRequest } from "next/server";
 import { z } from "zod";
 import { EventService } from "../services/event.service";
 import { createEventSchema, updateEventSchema } from "@/lib/validations";
+import {
+  BadRequestException,
+  ValidationException,
+} from "../utils/http-exception";
 
 class EventController {
   async createEvent(request: NextRequest) {
-    try {
-      const payload = await request.json();
-      const { googleAccessToken, ...eventData } = payload;
-      const body = createEventSchema.parse(eventData);
-      return await EventService.create(request, body, googleAccessToken);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return NextResponse.json(
-          { message: "Invalid request body", errors: error.errors },
-          { status: 400 }
-        );
-      }
-      throw error;
-    }
+    const payload = await request.json();
+
+    const { googleAccessToken, ...eventData } = payload;
+
+    const body = createEventSchema.parse(eventData);
+
+    return await EventService.create(request, body, googleAccessToken);
   }
 
   async getEvents(request: NextRequest) {
@@ -26,30 +23,35 @@ class EventController {
   }
 
   async getEvent(id: string) {
+    if (!id) {
+      throw new BadRequestException("Event ID is required");
+    }
     return await EventService.findById(id);
   }
 
-  async getEventBySlug(request: NextRequest, slug: string) {
+  async getEventBySlug(slug: string) {
+    if (!slug) {
+      throw new BadRequestException("Event Slug is required");
+    }
     return await EventService.findBySlug(slug);
   }
 
   async updateEvent(request: NextRequest, id: string) {
-    try {
-      const payload = await request.json();
-      const body = updateEventSchema.parse(payload);
-      return await EventService.update(request, id, body);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return NextResponse.json(
-          { message: "Invalid request body", errors: error.errors },
-          { status: 400 }
-        );
-      }
-      throw error;
+    const payload = await request.json();
+
+    if (!id) {
+      throw new BadRequestException("Event ID is required");
     }
+
+    const body = updateEventSchema.parse(payload);
+
+    return await EventService.update(request, id, body);
   }
 
   async deleteEvent(request: NextRequest, id: string) {
+    if (!id) {
+      throw new BadRequestException("Event ID is required");
+    }
     return await EventService.remove(request, id);
   }
 }
