+++ nebula-app/src/app/api/controllers/program.controller.ts
@@ -1,5 +1,6 @@
 import { NextRequest } from "next/server";
 import { ProgramService } from "../services/program.service";
+import { BadRequestException } from "../utils/http-exception";
 
 const programService = new ProgramService();
 
@@ -17,14 +18,23 @@ export class ProgramController {
   }
 
   async getBySlug(request: NextRequest, slug: string) {
+    if (!slug) {
+      throw new BadRequestException("Program Slug is required");
+    }
     return await programService.getBySlug(request, slug);
   }
 
   async updateProgram(request: NextRequest, slug: string) {
+    if (!slug) {
+      throw new BadRequestException("Program Slug is required");
+    }
     return await programService.updateProgram(request, slug);
   }
 
   async deleteProgram(request: NextRequest, slug: string) {
+    if (!slug) {
+      throw new BadRequestException("Program Slug is required");
+    }
     return await programService.deleteProgram(request, slug);
   }
 }
